requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
croniter==6.0.0
cryptography==46.0.5
databricks-sdk==0.33.0
databricks-sql-connector==3.4.0
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.17.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==6.6.0
protobuf==5.29.6
pyarrow==16.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytest==8.3.4
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.38.0
tenacity==8.5.0
thrift==0.20.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
